+++ src/ts/Native/Api/AR.ts
@@ -1,7 +1,9 @@
 import { ApiPackage, NativeApi } from 'Native/NativeApi';
 import { NativeBridge } from 'Native/NativeBridge';
 import { Observable0, Observable1, Observable2 } from 'Utilities/Observable';
-import { ARUtil, IARFrameInfo, IARFrameScale, IARPoint, IARRect, IARSize } from 'Utilities/ARUtil';
+import { IARSize } from 'Utilities/ARUtil';
+import { AndroidARApi } from 'Native/Api/AndroidARApi';
+import { IosARApi } from 'Native/Api/IosARApi';
 import { Platform } from 'Constants/Platform';
 
 enum AREvent {
@@ -16,16 +18,22 @@ enum AREvent {
     AR_SESSION_INTERRUPTION_ENDED
 }
 
-export interface IARConfigurationProperties {
+export interface IIosARConfigurationProperties {
     configurationName: string;
     lightEstimationEnabled?: boolean;
     worldAlignment?: number;
     planeDetection?: number;
 }
 
-export interface IARRunProperties {
+export interface IIosARRunProperties {
     runOptions?: number;
-    configuration?: IARConfigurationProperties;
+    configuration?: IIosARConfigurationProperties;
+}
+
+export interface IAndroidARRunProperties {
+    lightEstimationMode?: string;
+    planeFindingMode?: string;
+    updateMode?: string;
 }
 
 export interface IARVideoFormat {
@@ -34,6 +42,9 @@ export interface IARVideoFormat {
 }
 
 export class ARApi extends NativeApi {
+    public Android: AndroidARApi;
+    public Ios: IosARApi;
+
     public readonly onPlanesAdded = new Observable1<string>();
     public readonly onPlanesRemoved = new Observable1<string>();
     public readonly onPlanesUpdated = new Observable1<string>();
@@ -43,17 +54,18 @@ export class ARApi extends NativeApi {
     public readonly onError = new Observable1<number>();
     public readonly onSessionInterrupted = new Observable0();
     public readonly onSessionInterruptionEnded = new Observable0();
-    public readonly onAndroidEnumsReceived = new Observable1<any>();
 
     constructor(nativeBridge: NativeBridge) {
         super(nativeBridge, 'AR', ApiPackage.AR);
+
+        if (nativeBridge.getPlatform() === Platform.ANDROID) {
+            this.Android = new AndroidARApi(nativeBridge);
+        } else if (nativeBridge.getPlatform() === Platform.IOS) {
+            this.Ios = new IosARApi(nativeBridge);
+        }
     }
 
-    public isARSupported(): Promise<boolean> {
-        return this._nativeBridge.invoke<boolean>(this._fullApiClassName, 'isARSupported', ['ARWorldTrackingConfiguration']);
-    }
-
-    public restartSession(arRunProperties: IARRunProperties): Promise<void> {
+    public restartSession(arRunProperties: IIosARRunProperties|IAndroidARRunProperties): Promise<void> {
         return this._nativeBridge.invoke<void>(this._fullApiClassName, 'restartSession', [arRunProperties]);
     }
 
@@ -81,43 +93,8 @@ export class ARApi extends NativeApi {
         return this._nativeBridge.invoke<void>(this._fullApiClassName, 'removeAnchor', [identifier]);
     }
 
-    public advanceFrame(): Promise<void> {
-        // We don't have scaling logic in Android at the moment.
-        if (this._nativeBridge.getPlatform() === Platform.ANDROID) {
-            return this._nativeBridge.invoke<void>(this._fullApiClassName, 'advanceFrame');
-        }
-
-        // Get frame info, calculate scaling and then call advanceFrame
-        return this.getFrameInfo().then((frameInfo) => {
-            const scale = ARUtil.calculateVideoScale(frameInfo);
-            return this.setFrameScaling(scale).then(
-                () => this._nativeBridge.invoke<void>(this._fullApiClassName, 'advanceFrame')).catch(
-                    (error) => this._nativeBridge.Sdk.logInfo('Cannot set scaling: ' + error));
-        }).catch((error) => this._nativeBridge.Sdk.logInfo('Cannot get frame info: ' + error));
-    }
-
-    public getFrameInfo(): Promise<IARFrameInfo> {
-        return this._nativeBridge.invoke<IARFrameInfo>(this._fullApiClassName, 'getFrameInfo');
-    }
-
-    public setFrameScaling(scale: IARFrameScale): Promise<void> {
-        return this._nativeBridge.invoke<void>(this._fullApiClassName, 'setFrameScaling', [scale]);
-    }
-
     public getSupportedVideoFormats(): Promise<IARVideoFormat[]> {
         return this._nativeBridge.invoke<IARVideoFormat[]>(this._fullApiClassName, 'getSupportedVideoFormats');
-    }
-
-    public initAR(): Promise<void> {
-        if (this._nativeBridge.getPlatform() === Platform.ANDROID) {
-            return this._nativeBridge.invoke<any>(this._fullApiClassName, 'getAndroidConfigEnums').then((enums) => {
-                this.onAndroidEnumsReceived.trigger(enums);
-
-                return Promise.resolve();
-            });
-        }
-
-        return Promise.resolve();
     }
 
     public handleEvent(event: string, parameters: any[]): void {