+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-# Pending
+# Staged
+
+* Remove force quit manager
+* Start the VPAID timeout timer to hide ad unit on bad ads
+
+# Tue Aug 21 06:56:04 UTC 2018
 
 * Fix display interstitial click area
+* Rename comet tracking loaded event
+* Fix Vast Endscreen click delay
+* Fix VPAID click and Endscreen
+* Add creativeId and targetStoreId in XPromo Operative Events
 
 # Fri Aug 17 11:23:03 UTC 2018